Don't setup bootstrap's LINT0 pin as ExtINT (PIC) delivery mode.
We already mask the entire PIC, and install dumb handlers for
its setup vectors in case of any spurious interrupts.

There's simply no need for letting the chip's INTR pin reach the
local CPU through LINT0.

+++ include/ioapic.h
@@ -191,7 +191,7 @@ struct ioapic_pin {
 	int pin;			/* which pin in this ioapic */
 };
 
-void ioapic_setup_isairq(uint8_t irq, int vector);
+void ioapic_setup_isairq(uint8_t irq, uint8_t vector);
 void ioapic_init(void);
 
 #endif /* _IOAPIC_H */
